requirements.txt
Django==3.2.15
djongo==1.3.6
pymongo==3.12.3
djangorestframework
